Filter fixes - Extracted filter operator and expression building logic - Updated tests for custom eFunc filter
